Update lmo-tippsavewertgesamt.php
Fehlermeldung nach Spieltag und Ticker speichern als Admin mit aktiviertem Tippspiel der Liga.
PHP Warning:  A non-numeric value encountered in ../lmo/addon/tipp/lmo-tippsavewertgesamt.php on line 153

+++ lmo/addon/tipp/lmo-tippsavewertgesamt.php
@@ -127,13 +127,13 @@ for($k = 0; $k < $anzligen; $k++) {
       // Nick
       $op3 = (int)substr($dum[1], 2)-1;
       // Spieltagsnummer
-      $op4 = substr($dum[1], 0, 2);
+      $op4 = (int)substr($dum[1], 0, 2);
       // PP
       if ($op4 == "TP") {
         $gef = 0;
         for($j = 0; $j < $anztipper && $gef == 0; $j++) {
           if ($tippernick[$j] == $op1) {
-            $tipppunkte[$j][$anzligenaus] += $dum[2];
+            $tipppunkte[$j][$anzligenaus] += (int)$dum[2];
             $gef = 1;
           }
         }
@@ -141,7 +141,7 @@ for($k = 0; $k < $anzligen; $k++) {
         $gef = 0;
         for($j = 0; $j < $anztipper && $gef == 0; $j++) {
           if ($tippernick[$j] == $op1) {
-            $spielegetipptges[$j][$anzligenaus] += $dum[2];
+            $spielegetipptges[$j][$anzligenaus] += (int)$dum[2];
             $gef = 1;
           }
         }
@@ -150,7 +150,7 @@ for($k = 0; $k < $anzligen; $k++) {
         $gef = 0;
         for($j = 0; $j < $anztipper && $gef == 0; $j++) {
           if ($tippernick[$j] == $op1) {
-            $punkte[$j][$anzligenaus][$artpkt] += $dum[2];
+            $punkte[$j][$anzligenaus][$artpkt] += (int)$dum[2];
             $gef = 1;
           }
         }
